+++ frontend/components/loaders/Circle/Circle.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const baseClass = 'kolide-circle-loader';
+
+const Circle = () => {
+  return (
+    <div className={baseClass} />
+  );
+};
+
+export default Circle;