+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Show loading spinner while newly added Host Details are saved
+
 * Show a generic computer icon when when referring to hosts with an unknown platform instead of the text "All"
 
 * Kolide will now warn on startup if there are database migrations not yet completed.